feat: move shared components into shared folder
