+++ src/app/dashboard/club/manage/calendar/page.tsx
@@ -1,5 +1,5 @@
 import ClubInfo from "@/components/dashboard/club/calendar/molecules/ClubInfo";
-import ClubCalendar from "@/components/dashboard/club/calendar/organisms/ClubCalendar";
+import ClubCalendar from "@/components/dashboard/shared/organisms/ClubCalendar";
 import NewScheduleForm from "@/components/dashboard/club/calendar/organisms/NewScheduleForm";
 
 export default function CalendarPage() {
