+++ src/components/dashboard/shared/organisms/ClubPictureItem.tsx
